[minfs] Enable LSan for tests
LSan exception was added for inspector-test, which is actually in
zircon/system/utest/inspector/. So remove the exception.

This unearthed a few leaks in several tests, mostly related to death
situations. Most of the leaks were fixed by moving code outside of the
death lambda. In on situation though (TakeBlockReservationsWithInodeReservationDies)
we were losing ownership of the transaction in the death clause. I
didn't have a good way to fix this, so I ended up keeping track of
the raw pointer to delete it afterwards.

Change-Id: I19d09686bcf05406c7916d4499eb1a968e572bd0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/458937
Testability-Review: Chris Suter <csuter@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Kostya Kortchinsky <kostyak@google.com>