[fs] Use STL locking, fix use-after-free.
The fs library previously used FBL locking because when it was written,
STL was not permitted at that layer. This prohibition has not been the
case for some time.

FBL locking can not be used in host builds and we have accumulated many
ifdefs around this. In addition, we plan to do more locking as we work
to make filesystems more parallel.

Standardizing all "fs" locking on the STL primitives means we can
compile on host or target and avoid problems with lock types not
matching.

The ifdefs around locks are kept to keep this patch only changes the
lock types, to reduce risk of landing. Some lock ifdefs will be removed
in a second pass.

The lock change revealed as use-after-free in
ManagedVfs::OnShutdownComplete where a lock was unlocked after the lock
ifself was freed. Fix the problem. This was revealed by asan which for
some reason was not triggering on the old types.

Change-Id: Ia6cad929d5ed0e9dd0d22fe3458e34df25f36ab3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485744
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Vikram Auradkar <auradkar@google.com>