[minfs] Fix use of Cobalt library to avoid thread-safety issues
Cobalt seems to require that the dispatcher you use is only used from
one thread and that any thread you call it from is either the thread
that's running the loop, or a thread that doesn't have it set as the
default dispatcher.

This CL also removes a "delete this" and switches to memory management
via the more usual std::unique_ptr.

Bug: 74396
Fixed: 73218
Change-Id: I9c2ae5ebe1ed816c2749df7f98124688a72bd77d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/514689
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Manali Bhutiyani <manalib@google.com>