[vfs] Remove ManagedVfs default constructor.
The PagedVfs functionality will be moving into ManagedVfs and PagedVfs
requires a dispatcher in the constructor. This removes the default
constructor and updates the callers.

To tame some out-of-control ifdefs, a typedef is provided for the
cross-platform functions that need to take a dispatcher on Fuchsia and
nothing on the host.

Change-Id: I7d1d9ce4624642597c123d3414be4bf45c17e910
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564136
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>