[minfs] Temporarily disable minfs resize tests on ASAN builder
FIDL is undergoing a wire format migration. As part of this, a
transformer is used to transform between the old and new wire
format before/after decode/encode. This slows down the minfs
resize test so that it keeps timing out.

This CL disables the test on ASAN builds until the migration is
complete.

Bug: 84697
Change-Id: I8ee4577bd0d5172934401d599ae5bea0774796c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581901
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>