[fvm] Tidy up the FVM tool a bit
Fixed some quirks with the command-line parsing (some would cause the
tool to crash).

In the process of doing this, I discovered that the ffx was invoking the
fvm tool incorrectly (to be fair, the fault lies with the fvm tool which
is a bit clunky), and it was passing arguments that would be ignored.

Bug: 62215
Change-Id: I92b8646214a4c3ef294feaf15369729221f4e09b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580603
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>