[volume] Count max partition size in stats.
Counts the maximum partition size in the free_shared_pool_bytes in the
filesystem statistics. Previously this included the total free FVM
space, some of which might not be usable by the given partition.

Update the appmgr cache clearing now that the reported free slices is
correct.

Bug: 85718
Change-Id: I9d79b8296a1aeef169ccbaac940cd08ed538cb99
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/580902
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>