[minfs] quieter stat when not on fvm
When not on fvm this warning prints on every stat, which in one of the
unlink tests means it prints on the order of 70k times. Make the message
debug, and also turns up the sleep in the unlink test so we don't call
it as often while waiting for the blocks to become free.

Change-Id: I0bb28d7193817547e66e3abb431f02f1c3a423f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638841
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>