[storage] move fs services out of svc
When filesystem are launched as processes, they populate an outgoing
directory with the filesystem root and a variety of services.
Originally, the services would be put in a /svc directory, to mimic how
they are constructed for components.

However, this isn't what the exported directory looks like when a
filesystem is actually launched as a component. Instead, exported
capabilities are put in a single, flat directory. Filesystems should
mostly look the same whether launched as a component or as a process, so
we move the filesystem services to the root of the outgoing directory
when launched as a process.

This makes it simpler to write tooling that can interact with a
filesystem instance agnostic to how it was initially launched.

Bug: 91577
Test: CQ
Change-Id: I51ede148f6394f1750c6470db88c8dd36fdbc446
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652323
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>