[minfs] Refactor dirty cache metrics
Add new AllReservationsAllocated function to allow ensuring that the
space for a given reservation has already been allocated, and assert
that this is the case.

Add fsck checks to validate the number of allocated inodes/blocks
against the total reported in the superblock, ensuring if an assertion
happens due to this case, the device will be reformatted on next boot.

Previously, AddDirtyBytes would check that the volume could be extended
by the required number of blocks if we did not have enough space
allocated for the reservation. However, the volume should already have
been extended by BeginTransaction/ContinueTransaction.

Test: fx test
fuchsia-pkg://fuchsia.com/minfs-fs-tests#meta/minfs-unit-test.cm

Change-Id: I3d4e9baeb8e760a3ff4e41070448535d65fa201b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679744
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>