[minfs] separate the vfs object from the minfs object
Minfs and the vfs are closely linked, but we need to be able to start
one without having the information we need to start the other in order
to componentize the filesystems. This breaks the vfs object out into a
Runner class, which owns the Minfs object once it's created.

Right now, the runner class just immediately creates the minfs object.
There are a couple of extra accessors over the version of the Runner
class that is in blobfs because minfs tests love running the minfs code
directly.

Fixed: 100846
Test: fx test //src/storage/minfs
Change-Id: I6fbf39fa63acbe43e869b82deac50893d4af9ed4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683138
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>